Rename module to hots-dialog-renderer.js
...since it now exports a single class (Renderer), instead of a
collection of methods.
