+++ chrome-ext/src/js/hots-dialog-renderer.js
