+++ chrome-ext/src/manifest.json
@@ -34,7 +34,7 @@
         "js/hots-dialog.js",
         "js/hots-dialog-builder.js",
         "js/hots-dialog-paster.js",
-        "js/hots-dialog-renderers.js",
+        "js/hots-dialog-renderer.js",
         "js/hots-dialog-util.js"
       ]
     },
