+++ chrome-ext/tests/test-dialog.html
@@ -21,7 +21,7 @@
   <script src="../src/js/hots-dialog.js"></script>
   <script src="../src/js/hots-dialog-builder.js"></script>
   <script src="../src/js/hots-dialog-paster.js"></script>
-  <script src="../src/js/hots-dialog-renderers.js"></script>
+  <script src="../src/js/hots-dialog-renderer.js"></script>
   <script src="../src/js/hots-dialog-util.js"></script>
   <script>
     'use strict';
