+++ chrome-ext/tests/test-insert-hero-table.html
@@ -15,7 +15,7 @@
   <script src="./js/mocks.js"></script>
   <script src="../src/js/decorate-hots-data.js"></script>
   <script src="../src/js/hots-dialog.js"></script>
-  <script src="../src/js/hots-dialog-renderers.js"></script>
+  <script src="../src/js/hots-dialog-renderer.js"></script>
   <script src="../src/js/hots-dialog-util.js"></script>
   <script>
     (async function main() {
