+++ chrome-ext/tests/test-insert-templates.html
@@ -28,7 +28,7 @@
   <script src="./js/mocks.js"></script>
   <script src="../src/js/decorate-hots-data.js"></script>
   <script src="../src/js/hots-dialog.js"></script>
-  <script src="../src/js/hots-dialog-renderers.js"></script>
+  <script src="../src/js/hots-dialog-renderer.js"></script>
   <script>
     (async () => {
       //Load hero data for testing
